cleanup to easier file structure and handling of data files
