+++ data/locations/location_.json
@@ -1,0 +1,140 @@
+{
+    "trackParts": [
+        {
+            "id": "10",
+            "type": "Bumper",
+            "aSide": [
+                "5"
+            ],
+            "name": "bumper_10"
+        },
+        {
+            "id": "11",
+            "type": "Bumper",
+            "bSide": [
+                "2"
+            ],
+            "name": "bumper_11"
+        },
+        {
+            "id": "12",
+            "type": "Bumper",
+            "bSide": [
+                "3"
+            ],
+            "name": "bumper_12"
+        },
+        {
+            "id": "13",
+            "type": "Bumper",
+            "aSide": [
+                "4"
+            ],
+            "name": "bumper_13"
+        },
+        {
+            "id" : "1",
+            "bSide": [
+                "21"
+            ],
+            "aSide": [
+                "20"
+            ],
+            "length": 1000.0,
+            "name": "rail_1",
+            "sawMovementAllowed": false,
+            "parkingAllowed": true
+        },
+        {
+            "id" : "2",
+            "bSide": [
+                "20"
+            ],
+            "aSide": [
+                "11"
+            ],
+            "length": 100.0,
+            "name": "rail_2",
+            "sawMovementAllowed": false,
+            "parkingAllowed": true
+        },
+        {
+            "id" : "3",
+            "bSide": [
+                "20"
+            ],
+            "aSide": [
+                "12"
+            ],
+            "length": 100.0,
+            "name": "rail_3",
+            "sawMovementAllowed": false,
+            "parkingAllowed": true
+        },
+        {
+            "id" : "4",
+            "bSide": [
+                "13"
+            ],
+            "aSide": [
+                "21"
+            ],
+            "length": 100.0,
+            "name": "rail_4",
+            "sawMovementAllowed": false,
+            "parkingAllowed": true
+        },
+        {
+            "id" : "5",
+            "bSide": [
+                "10"
+            ],
+            "aSide": [
+                "21"
+            ],
+            "length": 100.0,
+            "name": "rail_5",
+            "sawMovementAllowed": false,
+            "parkingAllowed": true
+        },
+        {
+            "id": "20",
+            "type": "Switch",
+            "bSide": [
+                "1"
+            ],
+            "aSide": [
+                "3",
+                "2"
+            ],
+            "name": "switch_20"
+        },
+        {
+            "id": "21",
+            "type": "Switch",
+            "aSide": [
+                "1"
+            ],
+            "bSide": [
+                "4",
+                "5"
+            ],
+            "name": "switch_21"
+        }
+    ],
+	"facilities": [
+		{
+			"id": "20",
+			"type": "Reinigingsperron",
+			"relatedTrackParts": [
+				"1"
+			],
+			"taskTypes": [
+				{
+					"other": "Reinigingsperron"
+				}
+			],
+			"simultaneousUsageCount": 1
+		}
+    ]
+}
