+++ data/scenario_configurations/config_train_cleaning_late.json
@@ -1,0 +1,64 @@
+{
+    "location": "",
+    "start_time": 0,
+    "end_time": 72000,
+    "track_ids_used": true,
+    "use_default_material": true,
+    "trains_given": true,
+    "perform_servicing": true,
+    "partial_matching_given": false,
+    "partial_plan_given": false,
+    "through_traffic_given": false,
+    "custom_servicing_tasks": [
+        {
+            "name": "clean_train",
+            "type": "Reinigingsperron",
+            "priority": 1,
+            "duration": 500,
+            "required_skills": ["inwendige_reiniging"]
+        }
+    ],
+    "custom_train_units": [
+        {
+            "id": "2422",
+            "type": "SLT-4",
+            "services": ["clean_train"]
+        },
+        {
+            "id": "2301",
+            "type": "SNG-3",
+            "services": []
+        }
+    ],
+    "custom_trains": [
+        {
+            "id": "11111",
+            "members": ["2422"],
+            "arrival_time": 1500,
+            "arrival_track": "4",
+            "side_track_part": "13"
+
+        },
+        {
+            "id": "33333",
+            "members": ["2301"],
+            "arrival_time": 1900,
+            "arrival_track": "5",
+            "side_track_part": "10"
+        },
+        {
+            "id": "33334",
+            "member_types": ["SNG-3"],
+            "departure_time": 2000,
+            "departure_track": "3",
+            "side_track_part": "12"
+        },
+        {
+            "id": "11112",
+            "member_types": ["SLT-4"],
+            "departure_time": 2300,
+            "departure_track": "2",
+            "side_track_part": "11"
+        }
+    ]
+}
