+++ data/scenario_configurations/random_config_distribution1.json
@@ -1,5 +1,5 @@
 {
-    "location": "kleineBinckhorst",
+    "location": "kleineBinckhorst_HIP_dump",
     "start_time": 0,
     "end_time": 9600,
     "use_default_material": true,
@@ -8,8 +8,8 @@
     "number_of_trains": 10,
     "train_unit_distribution": {
         "train_unit_types": ["SLT-6", "SLT-4", "VIRM-4", "VIRM-6"],
-        "type_ratio": 0.5,
-        "units_per_composition": [3, 2],
+        "type_ratio": 0,
+        "units_per_composition": [3],
         "matching_complexity": 0
     }
 }
