--- src/scenario_generator/examples/enkhuizen/scenario_example_enkhuizen_solution.json
@@ -1,98 +1,0 @@
-[
-    {
-        "task": "move",
-        "train": "2000",
-        "path": ["21", "18", "16", "15", "1"],
-        "time": 600
-    },
-    {
-        "task": "shortstop",
-        "train": "2000",
-        "track": "1",
-        "time": 700
-    },
-    {
-        "task": "move",
-        "train": "2000",
-        "path": ["1", "15", "16", "20", "4"],
-        "time": 800
-    },
-    {
-        "task": "park",
-        "train": "2000",
-        "track" : "4",
-        "time": 900
-    },
-    {
-        "task": "inwendige_reiniging",
-        "train": "2000",
-        "track" : "4",
-        "time": 900
-    },
-    {
-        "task": "finish_serivce",
-        "train": "2000",
-        "track" : "4",
-        "time": 1800
-    },
-    {
-        "task": "move",
-        "train": "3000",
-        "path": ["21", "18", "16", "15", "2"],
-        "time": 900
-    },
-    {
-        "task": "shortstop",
-        "train": "3000",
-        "track": "2",
-        "time": 1000
-    },
-    {
-        "task": "move",
-        "train": "3000",
-        "path": ["2", "15", "16", "20", "7"],
-        "time": 1100
-    },
-    {
-        "task": "park",
-        "train": "3000",
-        "track" : "7",
-        "time": 1200
-    },
-    {
-        "task": "inwendige_reiniging",
-        "train": "3000",
-        "track" : "7",
-        "time": 1800
-    },
-    {
-        "task": "finish_serivce",
-        "train": "3000",
-        "track" : "7",
-        "time": 3600
-    },
-    {
-        "task": "move",
-        "train": "4001",
-        "path": ["5", "17", "16", "15", "2"],
-        "time": 1200
-    },
-    {
-        "task": "move",
-        "train": "4002",
-        "path": ["6", "17", "16", "15", "2"],
-        "time": 1300
-    },
-    {
-        "task": "combine",
-        "trains": ["4001", "4002"],
-        "track": "2",
-        "time": 1400
-    },    
-    {
-        "task": "move",
-        "train": "4000",
-        "path": ["2", "15", "16", "18", "21"],
-        "time": 1500
-    }
-]
